\name{MetaSubtract-package}
\alias{MetaSubtract-package}
\alias{MetaSubtract}
\docType{package}
\title{
\packageTitle{MetaSubtract}
}
\description{
  \packageDescription{MetaSubtract}
}
\details{
The core of this package is the function \code{\link{meta.subtract}}.
}
\author{
\packageAuthor{MetaSubtract}

Maintainer: \packageMaintainer{MetaSubtract}
}
\references{
Nolte et al. (2017). Missing heritability: is the gap closing? An analysis of 32 complex traits in the Lifelines Cohort Study. Eur J Hum Genet. 2017;25:877-885.
}
\seealso{ 
 \code{\link{meta.subtract}}.
}
\keyword{ package }
